renamed midi_socket into pipo_socket
